OpsManager separado en MatricesManager y VectoresManager para combinar los menus de vectores y matrices en uno solo; estructura de proyecto actualizada
