+++ gauss_bot/matrices/vector.py
